+++ server/tests-py/test_openapi.py
@@ -35,6 +35,9 @@ class TestOpenAPISpec:
     def test_multiple_endpoints_same_path(self, hge_ctx, transport):
         check_query_f(hge_ctx, self.dir() + '/openapi_multiple_endpoints_same_path.yaml', transport)
 
+    def test_multiple_endpoints_with_path_segments(self, hge_ctx, transport):
+        check_query_f(hge_ctx, self.dir() + '/openapi_multiple_endpoints_with_path_segments.yaml', transport)
+
     def test_endpoint_with_complex_arg(self, hge_ctx, transport):
         check_query_f(hge_ctx, self.dir() + '/openapi_get_endpoint_test_complex_arg.yaml', transport)