+++ CHANGELOG.md
@@ -31,6 +31,7 @@ count (
 - server: extend support for the `min`/`max` aggregates to all comparable types in BigQuery
 - server: fix support for joins in aggregates nodes in BigQuery
 - server: fix for passing input objects in query variables to remote schemas with type name customization (#7977)
+- server: fix REST endpoints with path segments not showing correctly in the OpenAPI spec
 - console: action/event trigger transforms are now called REST connectors
 - console: fix list of tables (and schemas) being unsorted when creating a new trigger event (fix #6391)
 - cli: migrate and seed subcommands has an option in prompt to choose and apply operation on all available databases